+++ docs/changelog.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## Unreleased
+## 0.9.14
 - Small tweak to the French (`"fr"`) localization.
+- Added Hebrew (`"he"`) localization (thank you @tinytim84!).
+- Added Thai (`"th"`) localization (thank you @samlaukinoon!).
 
 ## 0.9.13
 - Improvement to the French (`"fr"`) localization.